@@ -56,33 +56,33 @@
     "access": "public"
   },
   "peerDependencies": {
-    "@angular/core": "^9.0.0",
-    "@ngrx/store": "^9.0.0",
-    "@ngrx/effects": "^9.0.0"
+    "@angular/core": "^10.0.9",
+    "@ngrx/store": "^10.0.0",
+    "@ngrx/effects": "^10.0.0"
   },
   "dependencies": {
-    "immer": "^6.0.0"
+    "immer": "^7.0.7"
   },
   "devDependencies": {
-    "@angular/core": "^9.0.0",
-    "@ngrx/effects": "^9.0.0",
-    "@ngrx/store": "^9.0.0",
-    "@types/jest": "^24.0.11",
-    "@typescript-eslint/eslint-plugin": "^1.7.0",
-    "@typescript-eslint/parser": "^1.7.0",
-    "cpy-cli": "^3.1.0",
-    "eslint": "^5.16.0",
-    "eslint-config-prettier": "^4.2.0",
-    "eslint-plugin-prettier": "^3.0.1",
-    "husky": "^2.1.0",
-    "jest": "^24.7.1",
-    "lint-staged": "^8.1.5",
-    "microbundle": "^0.11.0",
-    "prettier": "^1.17.0",
+    "@angular/core": "^10.0.9",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@types/jest": "^26.0.9",
+    "@typescript-eslint/eslint-plugin": "^3.9.0",
+    "@typescript-eslint/parser": "^3.9.0",
+    "cpy-cli": "^3.1.1",
+    "eslint": "^7.6.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.2.5",
+    "jest": "^26.4.0",
+    "lint-staged": "^10.2.11",
+    "microbundle": "^0.12.3",
+    "prettier": "^2.0.5",
     "rimraf": "^3.0.2",
-    "semantic-release": "^15.13.3",
-    "ts-jest": "^24.0.2",
-    "tslint": "^5.0.0",
-    "typescript": "~3.7.4"
+    "semantic-release": "^17.1.1",
+    "ts-jest": "^26.2.0",
+    "tslint": "^6.1.3",
+    "typescript": "~3.9.7"
   }
 }
