@@ -3,8 +3,8 @@ import { ActionReducer, Action, On } from '@ngrx/store'
 
 export function createMutableReducer<S, A extends Action = Action, D = Draft<S>>(initialState: D, ...ons: On<D>[]) {
   const map = new Map<string, ActionReducer<D, A>>()
-  for (let on of ons) {
-    for (let type of on.types) {
+  for (const on of ons) {
+    for (const type of on.types) {
       if (map.has(type)) {
         const existingReducer = map.get(type) as ActionReducer<D, A>
         const newReducer: ActionReducer<D, A> = (state, action) => on.reducer(existingReducer(state, action), action)
@@ -15,7 +15,7 @@ export function createMutableReducer<S, A extends Action = Action, D = Draft<S>>
     }
   }
 
-  return function(state: D = initialState, action: A) {
+  return function (state: D = initialState, action: A) {
     return produce(state, (draft: D): D | undefined => {
       const reducer = map.get(action.type)
       return reducer ? reducer(draft, action) : undefined
